zoomIn() complete for now...

+++ Code/ComplexPlane.cpp
@@ -22,8 +22,7 @@ void ComplexPlane::zoomIn()
     m_ZoomCount++;
     float xSize = BASE_WIDTH * pow(BASE_ZOOM, m_ZoomCount);
     float ySize = BASE_HEIGHT * m_aspectRatio * pow(BASE_ZOOM, m_ZoomCount);
-    //m_plane_size(xSize, ySize); This doesn't work
-    m_plane_size = sf::Vector2f(static_cast<float>(xSize), static_cast<float>(ySize));
+    m_plane_size = sf::Vector2f(xSize, ySize);
     m_state = State::CALCULATING;
 }
